back/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.8.0
Pillow==10.1.0
PyJWT==2.8.0 